requirements.txt
jupyterlab==4.4.1
langchain==0.3.24
langchain_community==0.3.22
langchain_core==0.3.56
langchain_mistralai==0.2.10
mistralai==1.7.0
pdf2image==1.17.0
Pillow==11.2.1
pytesseract==0.3.13
python-dotenv==1.1.0
streamlit==1.44.1
unstructured[all-docs]==0.17.2
faiss-cpu==1.10.0
hf_xet==1.0.5
PyMuPDF==1.25.5